now when loggin with an unveified account it retruns a 423 and the UserStorage returns a byte size

+++ src/main/java/fr/nil/backedflow/responses/UserStorageResponse.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class UserStorageResponse {
 
-    private Float maxStorage;
-    private Float usedStorage;
+    private long maxStorage;
+    private long usedStorage;
 
 }
